Rename comps
