--- src/Wizard/WizardStepper/WizardStepperStep/index.js
@@ -1,1 +1,0 @@
-export { default } from "./WizardStepperStep";
