--- src/Wizard/WizardStepper/index.js
@@ -1,2 +1,0 @@
-export { default } from "./WizardStepper";
-export { default as WizardStepperStep } from "./WizardStepperStep";
