+++ src/Wizard/WizardSteps/WizardStep/index.js
@@ -1,0 +1,1 @@
+export { default } from "./WizardStep";
