+++ src/Wizard/WizardSteps/index.js
@@ -1,0 +1,2 @@
+export { default } from "./WizardSteps";
+export { default as WizardStep } from "./WizardStep";
